refactor: chat-app call modals using api

+++ apps/chat/src/features/ChatControl/ui/DeleteChat/DeleteChat.tsx
@@ -1,11 +1,12 @@
 import type { FC } from 'react';
-import { useState } from 'react';
 
-import { Button, Popup } from '@ducks-tinder-client/ui';
+import { Button, useOpenModal } from '@ducks-tinder-client/ui';
 
 import * as styles from './DeleteChat.module.scss';
 import { useTranslation } from 'react-i18next';
 import { useChatDispatch } from '@shared/lib/hooks';
+import type { ConfirmDeleteChatResult } from './ui';
+import { ConfirmDeleteChat } from './ui';
 import { deleteChatThunk } from '@entities/chat';
 
 interface DeleteChatProps {
@@ -15,47 +16,28 @@ interface DeleteChatProps {
 export const DeleteChat: FC<DeleteChatProps> = ({ submitDelete }) => {
   const { t } = useTranslation('chat');
 
+  const { openModal } = useOpenModal();
+
   const dispatch = useChatDispatch();
 
-  const [isChatDeleting, setIsChatDeleting] = useState(false);
+  const handleDeleteChat = async () => {
+    const confirmed = await openModal<never, ConfirmDeleteChatResult>({
+      Component: ConfirmDeleteChat,
+    });
 
-  const handleDelete = () => {
+    if (!confirmed) {
+      return;
+    }
+
     dispatch(deleteChatThunk());
     submitDelete?.();
   };
 
   return (
     <>
-      <Button
-        onClick={() => setIsChatDeleting(true)}
-        extraClassName={styles.btn}
-      >
+      <Button onClick={handleDeleteChat} extraClassName={styles.btn}>
         {t('delete')}
       </Button>
-      {isChatDeleting && (
-        <Popup
-          extraClassName={styles.overflow}
-          closeHandler={() => setIsChatDeleting(false)}
-          size="s"
-        >
-          <div className={styles.title}>{t('delete')}</div>
-          <div className={styles.descr}>
-            <div className={styles.text}>{t('sureWantToDelete')}</div>
-            <div className={styles.attention}>{t('noRecover')}</div>
-          </div>
-          <div className={styles.wrapper}>
-            <Button
-              onClick={() => setIsChatDeleting(false)}
-              extraClassName={styles.btn}
-            >
-              {t('cancel')}
-            </Button>
-            <Button onClick={handleDelete} extraClassName={styles.btn}>
-              {t('delete')}
-            </Button>
-          </div>
-        </Popup>
-      )}
     </>
   );
 };
